Refactor Code According to SonarQube
Changes mainly regard exception handling.
Any exceptions regarding the correctness of the proof have been switched
to checked exceptions.

+++ src/main/java/proof/validator/MainValidator.java
@@ -41,8 +41,8 @@ public class MainValidator implements Validator<JSONObject> {
           * graph.getNumberOfNodes()) {
         valid = true;
         Config.get().logger.progress("bound follows from Euler's polyhedron formula");
-      } else if (graph.getClaimedLowerBound() == Math.ceil(Math.pow(graph.getNumberOfEdges(), 3)
-          / (33.75 * Math.pow(graph.getNumberOfNodes(), 2)) - 3 * graph.getNumberOfNodes())) {
+      } else if (graph.getClaimedLowerBound() == (int) Math.ceil(Math.pow(graph.getNumberOfEdges(),
+          3) / (33.75 * Math.pow(graph.getNumberOfNodes(), 2)) - 3 * graph.getNumberOfNodes())) {
         valid = true;
         Config.get().logger.progress("bound follows from the formula by Pach & Todt");
       }
